Introduce AnyCost Stream example script
The example script is meant to use a very simple style.
A light number of Python constructs are used, a few
Python Standard Library modules are used for things
like reading and writing data files as well as safely
calculating discount cost. The only 3rd party library
is requests which we use as the interface is much more
readable and the Python Standard Library even advises
users to use it over the library facilities.